+++ yaki_admin/src/features/customer/components/ManageTeams
@@ -1,0 +1,73 @@
+<script setup="ts">
+  import router from "../../../../src/router/router";
+  import modalValidationState from "@/features/shared/services/modalValidationState";
+  import ModalValidation from "@/features/shared/components/ModalValidation.vue";
+  import { onBeforeMount } from "vue";
+  import { useTeamStore } from "../../../../src/stores/teamStore.js";
+  const teamStore = useTeamStore();
+  const fetchTeams = async () => {
+    await teamStore.getAllTeamsByCustomerId(teamStore.getCustomerId);
+  };
+  onBeforeMount(async () => {
+    fetchTeams();
+  });
+  const removeTeamFromCustomer = (id: number, informations: string) => {
+    teamStore.setTeamToDelete(id);
+    modalValidationState.setInformation(informations);
+    modalValidationState.changeVisibility();
+  };
+  const validationModalAccept = () => {
+    teamStore.deleteTeam(teamStore.getTeamToDelete);
+    setTimeout(() => {
+      fetchTeams();
+    }, 150);
+  };
+</script>
+<template>
+  <modal-validation
+    v-show="modalValidationState.isShowed"
+    @modal-accept="validationModalAccept" />
+  <div class="customer-view-teams">
+    <h1 class="title">Customer</h1>
+    <h2 class="text">Manage your teams</h2>
+    <hr class="line" />
+    <side-bar-button
+      v-bind:inner-text="'Add teams'"
+      v-bind:icon-path="plusIcon"
+      @click.prevent="router.push({ path: `invitation` })" />
+    <div class="team-list">
+      <team
+        :team="team"
+        v-for="team in teamStore.getTeamList"
+        :key="team.id"
+        @RemoteTeam="removeTeamFromCustomer" />
+    </div>
+  </div>
+</template>
+<style lang="scss">
+  @import url("https://fonts.googleapis.com/css2?family=Inter:wght@300&display=swap");
+  .customer-view-teams {
+    padding: 30px;
+    font-family: "Inter", sans-serif;
+  }
+  .title {
+    font-size: 38px;
+  }
+  .text {
+    font-size: 18px;
+    color: #787878;
+    margin-bottom: 20px;
+  }
+  .line {
+    width: 80%;
+    background-color: #efefefed;
+    margin-bottom: 1rem;
+  }
+  .team-list {
+    padding: 50px;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    gap: 3rem;
+  }
+</style>
